+++ NEWS.md
@@ -1,5 +1,8 @@
 # forcats 0.1.1.9000
 
+* New `fct_other()` makes it easier to convert selected levels to "other" 
+  (#40)
+
 * New `fct_relabel()` allows relabeling factor levels programmatically (#50, @krlmlr).
 
 * `fct_drop()` gains `only` argument to restrict which levels are dropped (#69)